Refactor T-D decoupling code as a submodule
